feat: improve backend code organization
