+++ src-tauri/Cargo.toml
@@ -17,29 +17,39 @@ tauri-build = { version = "1.5", features = [] }
 serde_json = "1.0.115"
 tauri = { version = "1.6", features = [ "updater", "api-all", "macos-private-api", "system-tray", "icon-png"] }
 tauri-plugin-positioner = { version = "1.0", features = ["system-tray"] }
-tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
+tokio = { version = "1.37.0", features = ["rt-multi-thread", "macros", "full"] }
 rusqlite = { version = "0.31.0", features = ["bundled"] }
 dirs = "5.0.1"
-kubeforward = { path = "./utils/kubeforward" }
 reqwest = "0.12.3"
 base64 = "0.22.0"
 keyring = "2.3.2"
 mockall = "0.12"
 hostsfile = { git = "https://github.com/tonarino/innernet", branch = "main" }
 device_query = "=2.0.0"
-
 log = "0.4"
 env_logger = "0.11.3"
 openssl = { version = "0.10.64", features = ["vendored"] }
 openssl-sys = {version = "0.9.102", features = ["vendored"] }
+kube = { version = "0.88.0", features = ["client", "config", "rustls-tls", "ws"] }
+k8s-openapi = { version = "0.21.0", default-features = false, features = ["latest"] }
+tokio-stream = { version = "0.1.15", features = ["net"] }
+futures = "0.3.30"
+anyhow = "1.0.81"
+tracing = "0.1.40"
+hyper = { version = "0.14.28", features = [ "client", "http1", "http2", "tcp", "stream" ] }
+lazy_static = "1.4.0"
+serde = { version = "1.0", features = ["derive"] }
+rand = "0.8.5"
+kube-runtime = "0.88.0"
+tower = "0.4.13"
+whoami = "1.5.1"
+tempfile = "3.9"
+h2 = { optional = true, version = "0.4.4" }
+fix-path-env = { git = "https://github.com/tauri-apps/fix-path-env-rs" }
 
 [dev-dependencies]
 tempfile = "3.9"
-
-[dependencies.h2]
-optional = true
-version = "0.4.4"
-
+tracing-subscriber = { version = "0.3.18", features = [ "env-filter" ] }
 
 [features]
 # by default Tauri runs in production mode
@@ -48,6 +58,3 @@ default = ["custom-protocol"]
 # this feature is used used for production builds where `devPath` points to the filesystem
 # DO NOT remove this
 custom-protocol = ["tauri/custom-protocol"]
-
-[dependencies.fix-path-env]
-git = "https://github.com/tauri-apps/fix-path-env-rs"
