+++ src-tauri/src/kubeforward/kubecontext.rs
@@ -1,4 +1,4 @@
-use crate::vx::Pod;
+use crate::kubeforward::vx::Pod;
 use anyhow::{Context, Result};
 use k8s_openapi::{
     api::core::v1::{Namespace, Service},
@@ -13,7 +13,10 @@ use kube::{
     config::{Config, KubeConfigOptions, Kubeconfig},
     Client, ResourceExt,
 };
-use serde::Serialize;
+
+use crate::models::kube::{
+    KubeContextInfo, KubeNamespaceInfo, KubeServiceInfo, KubeServicePortInfo,
+};
 
 pub async fn create_client_with_specific_context(
     kubeconfig_path: Option<&Path>,
@@ -47,27 +50,6 @@ pub async fn create_client_with_specific_context(
     let client = Client::new(service, config.default_namespace);
 
     Ok(client)
-}
-
-#[derive(Serialize)]
-pub struct KubeContextInfo {
-    pub name: String,
-}
-
-#[derive(Serialize)]
-pub struct KubeNamespaceInfo {
-    pub name: String,
-}
-
-#[derive(Serialize)]
-pub struct KubeServiceInfo {
-    pub name: String,
-}
-
-#[derive(Serialize)]
-pub struct KubeServicePortInfo {
-    pub name: Option<String>,
-    pub port: Option<IntOrString>,
 }
 
 #[tauri::command]
@@ -159,27 +141,27 @@ pub async fn list_service_ports(
         if let Some(service_ports) = spec.ports {
             for sp in service_ports {
                 if let Some(IntOrString::String(ref name)) = sp.target_port {
-                    let selector_string = spec.selector.as_ref().map_or_else(
-                        || String::new(),
-                        |s| {
-                            s.iter()
-                                .map(|(key, value)| format!("{}={}", key, value))
-                                .collect::<Vec<_>>()
-                                .join(", ")
-                        },
-                    );
+                    // Construct a selector string from the pod's labels, if available.
+                    let selector_string = spec.selector.as_ref().map_or_else(String::new, |s| {
+                        s.iter()
+                            .map(|(key, value)| format!("{}={}", key, value))
+                            .collect::<Vec<_>>()
+                            .join(", ")
+                    });
 
+                    // Attempt to list pods using the constructed label selector.
                     let pods = api_pod
-                        .list(&ListParams::default().labels(&selector_string))
+                        .list(&ListParams::default().labels(selector_string.as_str())) // Correctly pass a &str
                         .await
                         .map_err(|e| format!("Failed to list pods: {}", e))?;
 
+                    // Iterate through the list of pods to find a matching container port name.
                     'port_search: for pod in pods {
                         if let Some(spec) = &pod.spec {
                             for container in &spec.containers {
                                 if let Some(ports) = &container.ports {
                                     for cp in ports {
-                                        // Match the port name
+                                        // Match the port name and add the port info to the service_port_infos vector if found.
                                         if cp.name.as_deref() == Some(name) {
                                             service_port_infos.push(KubeServicePortInfo {
                                                 name: cp.name.clone(),
