+++ src-tauri/src/kubeforward/proxy.rs
@@ -1,5 +1,5 @@
 #![allow(clippy::too_many_arguments)]
-use crate::kubecontext::create_client_with_specific_context;
+use crate::kubeforward::kubecontext::create_client_with_specific_context;
 use k8s_openapi::api::core::v1::Pod;
 use kube::api::{DeleteParams, ListParams};
 use kube::{api::Api, Client};
@@ -11,9 +11,11 @@ use std::io::Read;
 use std::path::PathBuf;
 use std::time::{SystemTime, UNIX_EPOCH};
 
-use crate::port_forward::{
-    start_port_forward, start_port_udp_forward, stop_port_forward, Config, CustomResponse,
+use crate::kubeforward::port_forward::{
+    start_port_forward, start_port_udp_forward, stop_port_forward,
 };
+use crate::models::config::Config;
+use crate::models::response::CustomResponse;
 
 fn get_pod_manifest_path() -> PathBuf {
     let home_dir = dirs::home_dir().expect("Failed to resolve home directory");
