+++ src-tauri/src/models/mod.rs
@@ -1,0 +1,5 @@
+// src/models/mod.rs
+pub mod config;
+pub mod dialog;
+pub mod kube;
+pub mod response;
