--- src-tauri/utils/kubeforward/Cargo.toml
@@ -1,37 +1,0 @@
-[package]
-name = "kubeforward"
-description = "A low-level Kubernetes forwarding library"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tauri = { version = "1.6", features = [ "api-all", "macos-private-api", "system-tray"] }
-kube = { version = "0.88.0", features = ["client", "config", "rustls-tls", "ws"] }
-k8s-openapi = { version = "0.21.0", default-features = false, features = ["latest"] }
-tokio = { version = "1.37.0", features = [ "full" ] }
-tokio-stream = { version = "0.1.15", features = ["net"] }
-futures = "0.3.30"
-anyhow = "1.0.81"
-tracing = "0.1.40"
-serde_json = "1.0.115"
-hyper = { version = "0.14.28", features = [ "client", "http1", "http2", "tcp", "stream" ] }
-lazy_static = "1.4.0"
-serde = { version = "1.0", features = ["derive"] }
-log = "0.4"
-rand = "0.8.5"
-kube-runtime = "0.88.0"
-tower = "0.4.13"
-whoami = "1.5.1"
-dirs = "5.0.1"
-tempfile = "3"
-hostsfile = { git = "https://github.com/tonarino/innernet", branch = "main" }
-
-
-[dependencies.h2]
-optional = true
-version = "0.4.4"
-
-[dev-dependencies]
-tracing-subscriber = { version = "0.3.18", features = [ "env-filter" ] }
