--- src-tauri/utils/kubeforward/src/.gitignore
@@ -1,5 +1,0 @@
-# Generated by Cargo
-# will have compiled files and executables
-/target/
-
-
